fix minor dist related issues
Add .mailmap for Miyagawa's new email address.

Add perl 5.8 to travis config, we need to make sure the %SIG
localization code passes on 5.8.8+ .

Use my ContributorsFromPod plugin to override ContributorsFromGit in
Milla, because we have contributors that are not in the Git history.

Set GithubMeta.user = miyagawa in case other people build the dist.

Rebuild dist with above changes.