improve exit code handling in #12 and misc.
Preserve the old behavior of throwing an exception on explicit non-zero
exit, unless this is overridden in the object or globally. Add tests.

Truncate exit value to 0-255 as per POSIX.

Fix a warning on perl 5.8.

Update .gitignore and Changes.

Allow passing arguments to the CGI coderef, set in both @_ and @ARGV
locally.

Update POD for modified return code behavior and the argument passing.

Remove 'no warnings' in the CGI, don't know how that got in there, but
the user should handle that herself. Perhaps we should add an option for
this as well.