fix signal related test fails on Win32 #16
Workaround various bugs in the Win32 perl.

Fix a bogus USR1 entry in %SIG causing warnings and a test fail. The
USR1 entry seems to both exist and throw a warning that it is not
supported.

As for local-SIG.t, there is no way to catch a QUIT or INT simulated
signal, it terminates the whole process, so we skip the test on Win32.