Fix race condition in temp dir handling.
Set the temp dir location on first usage instead of during compile
time, this prevents a race condition if the process forks after module
compilation.

Also only try to remove the temp dir if the variable has been set, it
will not be if no CGI has been compiled.

This race condition is tested for in t/race-conditions.t, added in
c0a01bc, which now passes.

Fix #24.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>