Use better package/sub names for CGI coderefs.
Fix #8.

Use path of CGI script as identifier-safe package name, and the filename
as the identifier-safe coderef name.

This replaces the previous behavior of using the full path as the
package name for an anon coderef.

Replace the private method:

$package = $self->_build_package($path);

with

($package, $subname) = $self->_build_subname($path);

and update the documentation.

For scalar refs, use the caller's package with a counter per-package, to
form names of the form __CGI0__, __CGI1__ etc.. Instead of regular
anonymous coderefs in CGI::Compile.

Add t/symbols.t to test the assigned package and coderef names in both
cases.

Add main dependency on Sub::Name and a test dependency on Sub::Identify.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>